Refactoring: Renaming File notification.py to notification_service.py
