+++ WebService/key_clip_service.py
@@ -5,7 +5,7 @@ from sqlalchemy.orm import sessionmaker, scoped_session
 from queue import Queue
 from Services.s3_service import S3Service
 from Models.events import Events
-from notification import Notification
+from notification_service import Notification
 import mysql.connector
 import time
 import cv2
